backend/requirements.txt
Django
Djangorestframework
drf-yasg
django-cors-headers
requests
psycopg2
